+++ CHANGELOG.md
@@ -270,3 +270,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Layout folder
 - Grid component
+- GridColumn component